requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alpaca-trade-api==3.2.0
anyio==3.6.2
APScheduler==3.10.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==24.2.0
backcall==0.2.0
backtrader
beautifulsoup4==4.12.3
blis==0.7.11
boto3==1.28.59
botocore==1.31.59
bs4==0.0.2
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpathlib==0.15.1
colorama==0.4.6
comm==0.1.3
confection==0.1.3
contourpy==1.3.0
cryptography==40.0.2
cycler==0.12.1
cymem==2.0.8
debugpy==1.6.7
decorator==5.1.1
deprecation==2.1.0
distro==1.8.0
docker==7.1.0
et-xmlfile==1.1.0
eventkit==1.0.0
eventregistry==9.1
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.88.0
Flask==2.2.2
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
greenlet==2.0.1
h11==0.14.0
ib-insync==0.9.80
ibeam==0.5.5
idna==3.4
importlib-metadata==5.1.0
importlib_resources==6.5.2
ipykernel==6.23.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.7
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.9.4
matplotlib-inline==0.1.6
msgpack==1.0.3
multidict==6.2.0
multitasking==0.0.11
murmurhash==1.0.10
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.23.5
openpyxl==3.1.2
outcome==1.3.0.post0
package-name==0.1
packaging==23.1
pandas==1.5.2
parso==0.8.3
pathy==0.10.2
pdfminer.six==20221105
pdfplumber==0.7.6
peewee==3.17.9
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
preshed==3.0.9
prompt-toolkit==3.0.38
propcache==0.3.1
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.15.1
PyMySQL==1.0.2
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv
pytz==2022.6
PyVirtualDisplay==3.0
# pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
redis==4.4.0
regex==2023.8.8
requests==2.31.0
s3transfer==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.12.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.45
srsly==2.4.8
stack-data==0.6.2
starlette==0.22.0
tabula-py==2.7.0
tabulate==0.9.0
thinc==8.2.1
threadpoolctl==3.6.0
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.26.2
trio-websocket==0.11.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.14
uvicorn==0.20.0
Wand==0.6.11
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.55
zipp==3.11.0
